+++ boomer.go
@@ -7,7 +7,7 @@ import (
 	"github.com/jinzhu/copier"
 	"github.com/rs/zerolog/log"
 
-	funcPlugin "github.com/httprunner/func-plugin"
+	"github.com/httprunner/funplugin"
 	"github.com/httprunner/hrp/internal/boomer"
 	"github.com/httprunner/hrp/internal/ga"
 )
@@ -22,8 +22,8 @@ func NewBoomer(spawnCount int, spawnRate float64) *HRPBoomer {
 
 type HRPBoomer struct {
 	*boomer.Boomer
-	plugins      []funcPlugin.IPlugin // each task has its own plugin process
-	pluginsMutex *sync.RWMutex        // avoid data race
+	plugins      []funplugin.IPlugin // each task has its own plugin process
+	pluginsMutex *sync.RWMutex       // avoid data race
 }
 
 // Run starts to run load test for one or multiple testcases.